+++ src/compas_fab/backends/vrep/backend_features/vrep_add_attached_collision_mesh.py
@@ -2,15 +2,17 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from compas_fab.backends.vrep.helpers import assert_robot
+from compas_fab.backends.interfaces import AddAttachedCollisionMesh
 from compas_fab.backends.vrep.helpers import DEFAULT_OP_MODE
 from compas_fab.backends.vrep.helpers import VrepError
-from compas_fab.backends.interfaces import AddAttachedCollisionMesh
-from compas_fab.backends.vrep.remote_api import vrep
+from compas_fab.backends.vrep.helpers import assert_robot
+from compas_fab.utilities import LazyLoader
 
 __all__ = [
     'VrepAddAttachedCollisionMesh',
 ]
+
+vrep = LazyLoader('vrep', globals(), 'compas_fab.backends.vrep.remote_api.vrep')
 
 
 class VrepAddAttachedCollisionMesh(AddAttachedCollisionMesh):