+++ src/compas_fab/ghpython/components/install.py
@@ -1,13 +1,12 @@
 import argparse
 import glob
 import os
-import shutil
 
+from compas._os import create_symlinks
+from compas._os import remove_symlinks
 from compas_ghpython import get_grasshopper_library_path
+from compas_ghpython import get_grasshopper_userobjects_path
 from compas_rhino import _check_rhino_version
-
-from compas_fab.ghpython.components import assembly_name
-from compas_fab.ghpython.components import get_ghpy_filename
 
 
 def get_version_from_args():
@@ -25,15 +24,24 @@ def install():
         version = get_version_from_args()
         grasshopper_library_path = get_grasshopper_library_path(version)
 
-        for f in _remove_old_components(grasshopper_library_path):
-            results.append(('compas_fab', 'Uninstalled existing component: {}'.format(f), True))
+        for f in _remove_old_ghpy_components(grasshopper_library_path):
+            results.append(('compas_fab', 'Uninstalled old GHPY component: {}'.format(f), True))
 
-        src = os.path.join(os.path.dirname(__file__), get_ghpy_filename())
-        dst = os.path.join(grasshopper_library_path, get_ghpy_filename())
+        dstdir = get_grasshopper_userobjects_path(version)
+        srcdir = os.path.dirname(__file__)
+        userobjects = glob.glob(os.path.join(srcdir, '*.ghuser'))
 
-        shutil.copyfile(src, dst)
+        symlinks_to_remove = []
+        symlinks_to_add = []
+        for src in userobjects:
+            dst = os.path.join(dstdir, os.path.basename(src))
+            symlinks_to_remove.append(dst)
+            symlinks_to_add.append((src, dst))
 
-        results.append(('compas_fab', 'GH Components installed on {}'.format(grasshopper_library_path), True))
+        remove_symlinks(symlinks_to_remove)
+        create_symlinks(symlinks_to_add)
+
+        results.append(('compas_fab', 'Installed {} GH User Objects'.format(len(userobjects)), True))
     except PermissionError:
         raise Exception('Please close first all instances of Rhino and then rerun the command')
 
@@ -48,16 +56,29 @@ def uninstall():
         version = get_version_from_args()
         grasshopper_library_path = get_grasshopper_library_path(version)
 
-        for f in _remove_old_components(grasshopper_library_path):
-            results.append(('compas_fab', 'Uninstalled existing component: {}'.format(f), True))
+        for f in _remove_old_ghpy_components(grasshopper_library_path):
+            results.append(('compas_fab', 'Uninstalled old component: {}'.format(f), True))
+
+        dstdir = get_grasshopper_userobjects_path(version)
+        srcdir = os.path.dirname(__file__)
+        userobjects = glob.glob(os.path.join(srcdir, '*.ghuser'))
+
+        symlinks = []
+        for src in userobjects:
+            dst = os.path.join(dstdir, os.path.basename(src))
+            symlinks.append(dst)
+
+        remove_symlinks(symlinks)
+
+        results.append(('compas_fab', 'Uninstalled {} GH User Objects'.format(len(userobjects)), True))
     except PermissionError:
         raise Exception('Please close first all instances of Rhino and then rerun the command')
 
     return results
 
 
-def _remove_old_components(grasshopper_library_path):
-    oldlibs = glob.glob('%s*.ghpy' % os.path.join(grasshopper_library_path, assembly_name))
+def _remove_old_ghpy_components(grasshopper_library_path):
+    oldlibs = glob.glob('%s*.ghpy' % os.path.join(grasshopper_library_path, 'COMPAS FAB_0.1'))
     for f in oldlibs:
         os.remove(f)
         yield f